Working loading of user information using decorated-ish classes
