+++ src/containers/SmartCompassApp.jsx
@@ -1,0 +1,5 @@
+import { CompassApp } from "./CompassApp";
+import withSmartData from './withSmartData';
+
+const SmartCompassApp = withSmartData(CompassApp);
+export default SmartCompassApp;
